src/users: add age and fullname properties; get properties; rename unit
Adding a few properties that were supposed to be added being age and
full name
Adding a procedure to save this information in the database.
Adding a function to check whether the user has logged in before.
Rename the unit to have a capital U at the end: `User_u` -> `User_U`